Construct block textures from elements with faces. Not completely correct yet. Readd showing individual block textures for debugging

+++ MCPixelArtConverter/MCPoint.cs
@@ -9,15 +9,15 @@ namespace MCPixelArtConverter
 {
     class MCPoint
     {
-        public Double x { get; }
-        public Double y { get; }
-        public Double z { get; }
+        public float x { get; }
+        public float y { get; }
+        public float z { get; }
 
         public MCPoint(JArray json)
         {
-            x = Double.Parse(json[0].ToString());
-            y = Double.Parse(json[1].ToString());
-            z = Double.Parse(json[2].ToString());
+            x = float.Parse(json[0].ToString());
+            y = float.Parse(json[1].ToString());
+            z = float.Parse(json[2].ToString());
         }
     }
 }
